Implement buzzing.
Test by flashing a script using uflash and the custom firmware.hex

Reusing some of the runtime functions (like sleep, set_digital).
Using a custom implemntation (even a literal copy of the respective code)
led to the scripts not being executed (no errors shown however).
This behaviour does not depend on the content of the script.